Use an Arbitrary_Object as default
