+++ includes/activity/class-activity.php
@@ -159,21 +159,12 @@ class Activity extends Base_Object {
 	 *
 	 * @see https://www.w3.org/TR/activitypub/#object-without-create
 	 *
-	 * @param array|string|Base_Object|Link|null $data         Activity object.
-	 * @param string|null                        $object_class The full class path to a child of Base_Object.
+	 * @param array|string|Base_Object|Link|null $data Activity object.
 	 */
-	public function set_object( $data, $object_class = null ) {
+	public function set_object( $data ) {
 		// Convert array to object.
 		if ( is_array( $data ) ) {
-			if ( $object_class && class_exists( $object_class ) ) {
-				$data = $object_class::init_from_array( $data );
-			} elseif ( is_activity( $data ) ) { // Check if the item is an Activity or an Object.
-				$data = self::init_from_array( $data );
-			} elseif ( is_actor( $data ) ) {
-				$data = Actor::init_from_array( $data );
-			} else {
-				$data = Base_Object::init_from_array( $data );
-			}
+			$data = Arbitrary_Object::init_from_array( $data );
 		}
 
 		// Set object.
