+++ includes/collection/class-outbox.php
@@ -73,12 +73,6 @@ class Outbox {
 			),
 		);
 
-		$object_class = get_class( $activity_object );
-
-		if ( Base_Object::class !== $object_class ) {
-			$outbox_item['meta_input']['_activitypub_object_class'] = str_replace( '\\', '\\\\', $object_class );
-		}
-
 		$has_kses = false !== \has_filter( 'content_save_pre', 'wp_filter_post_kses' );
 		if ( $has_kses ) {
 			// Prevent KSES from corrupting JSON in post_content.
